# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pyn-server",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/admin/limits": {
      "get": {
        "tags": [
          "limits_api"
        ],
        "operationId": "list_limits",
        "responses": {
          "200": {
            "description": "the server default and every owner with limits of its own",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitsListing"
                }
              }
            }
          },
          "403": {
            "description": "server_admin_required, or service_scope_required for a service credential without manage_limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/owners/{owner}/limits": {
      "patch": {
        "tags": [
          "limits_api"
        ],
        "operationId": "set_owner_limits",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "description": "a user or organization name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLimitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "lowering a limit below current use deletes nothing; it only stops growth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerLimits"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request: a value out of range, or a member limit on a user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "server_admin_required, or service_scope_required for a service credential without manage_limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "user_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/owners/{owner}/limits": {
      "get": {
        "tags": [
          "limits_api"
        ],
        "operationId": "owner_limits",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "description": "a user or organization name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all limits null means unlimited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerLimits"
                }
              }
            }
          },
          "403": {
            "description": "not_namespace_owner or not_org_owner: only the owner, an administrator or a service credential with manage_limits may look",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "user_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/owners/{owner}/usage": {
      "get": {
        "tags": [
          "limits_api"
        ],
        "operationId": "owner_usage",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "description": "a user or organization name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "usage next to the limits in force, with every repository the owner has",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerUsage"
                }
              }
            }
          },
          "403": {
            "description": "not_namespace_owner or not_org_owner: only the owner, an administrator or a service credential with manage_limits may look",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "user_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/repos/{owner}/{name}/usage": {
      "get": {
        "tags": [
          "limits_api"
        ],
        "operationId": "repo_usage",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "description": "The namespace that owns the repository.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "The repository's name within its owner.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "readable by anyone who can read the repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoUsage"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/session": {
      "get": {
        "tags": [
          "session_api"
        ],
        "operationId": "current",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "401": {
            "description": "not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "session_api"
        ],
        "operationId": "sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "sets the HttpOnly session cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "right password, but the account cannot sign in: email_not_verified, approval_pending or account_disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "too_many_attempts; Retry-After says how many seconds to wait",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "session_api"
        ],
        "operationId": "sign_out",
        "responses": {
          "204": {
            "description": "clears the cookie"
          },
          "403": {
            "description": "csrf_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Body of every non-2xx response; `code` is stable and machine-readable.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Limits": {
        "type": "object",
        "description": "Limits on one owner. Every limit is opt-in: a null field is unlimited.",
        "required": [
          "repositories",
          "members",
          "storage_bytes"
        ],
        "properties": {
          "members": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Members an organization may have; always null for a user.",
            "minimum": 0
          },
          "repositories": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Repositories the owner may have.",
            "minimum": 0
          },
          "storage_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes of distinct content across the owner's repositories.",
            "minimum": 0
          }
        }
      },
      "LimitsListing": {
        "type": "object",
        "description": "The server default and every owner that has limits of its own.",
        "required": [
          "defaults",
          "owners"
        ],
        "properties": {
          "defaults": {
            "$ref": "#/components/schemas/Limits",
            "description": "Applies to owners with nothing of their own; all null unless the operator configures it."
          },
          "owners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnerLimitsEntry"
            },
            "description": "Ordered by name."
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "OwnerKind": {
        "type": "string",
        "description": "What an owner (a user or an organization) is.",
        "enum": [
          "user",
          "org"
        ]
      },
      "OwnerLimits": {
        "type": "object",
        "description": "An owner's limits. Nothing is limited unless an operator sets a limit or a server default.",
        "required": [
          "kind",
          "effective",
          "own"
        ],
        "properties": {
          "effective": {
            "$ref": "#/components/schemas/Limits",
            "description": "The limits in force: the owner's own where set, else the server default, else null (unlimited)."
          },
          "kind": {
            "$ref": "#/components/schemas/OwnerKind"
          },
          "own": {
            "$ref": "#/components/schemas/Limits",
            "description": "Limits set for this owner alone; a null field follows the server default."
          }
        }
      },
      "OwnerLimitsEntry": {
        "type": "object",
        "required": [
          "owner",
          "limits"
        ],
        "properties": {
          "limits": {
            "$ref": "#/components/schemas/OwnerLimits"
          },
          "owner": {
            "type": "string"
          }
        }
      },
      "OwnerUsage": {
        "type": "object",
        "description": "What an owner uses against its limits.",
        "required": [
          "owner",
          "limits",
          "usage",
          "repositories"
        ],
        "properties": {
          "limits": {
            "$ref": "#/components/schemas/OwnerLimits"
          },
          "owner": {
            "type": "string"
          },
          "repositories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepoUsage"
            },
            "description": "Ordered by name."
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        }
      },
...
      "RepoUsage": {
        "type": "object",
        "required": [
          "repository",
          "stored_bytes",
          "files",
          "revisions"
        ],
        "properties": {
          "files": {
            "type": "integer",
            "format": "int64",
            "description": "Paths with a head revision.",
            "minimum": 0
          },
          "repository": {
            "type": "string",
            "description": "`owner/name`."
          },
          "revisions": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "stored_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "SessionInfo": {
        "type": "object",
        "required": [
          "user",
          "csrf_token",
          "expires_at"
        ],
        "properties": {
          "csrf_token": {
            "type": "string",
            "description": "Send as the `X-Pyn-CSRF` header on every POST, PUT and DELETE."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "string"
          }
        }
      },
...
      "SetLimitsRequest": {
        "type": "object",
        "description": "Fields left out stay as they are; a number sets the limit and `null` returns the field to the server default.",
        "properties": {
          "members": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Organizations only; a user has no member limit.",
            "minimum": 0
          },
          "repositories": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "storage_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "Usage": {
        "type": "object",
        "required": [
          "repositories",
          "members",
          "stored_bytes"
        ],
        "properties": {
          "members": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Organization members; null for a user.",
            "minimum": 0
          },
          "repositories": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "stored_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Bytes of distinct content, counted once per repository.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}